display: Define soong namespace
HALs are starting to move to blueprint files, which means that pathmap
ifeqs no longer work. This results in modules with the same name in
different platform HAL directories. Since these modules aren't used or
linked anywhere outside the HAL directories and a device's device.mk,
we can safely use a soong namespace to isolate the different HAL
directories' blueprints.

Change-Id: Ibe275bd3183af68c36a516d24f3792ec520e60bc
Signed-off-by: Harsh Shandilya <harsh@prjkt.io>